refactor: Made test utility package
Created an internal test utility package which contains utility
functions for tests. This should not show up in the final binary
according to https://golang.org/doc/go1.4#internalpackages

This should make it easier to create new tests and improve reuse.

Contributes to: #18
