+++ framework/models/procedure_test.go
@@ -4,12 +4,10 @@ import (
 	"fmt"
 	"io/ioutil"
 	"os"
-	"path"
-	"path/filepath"
-	"runtime"
 	"testing"
 
 	"github.com/go-yaml/yaml"
+	"github.com/julianGoh17/simple-e2e/framework/internal"
 	"github.com/julianGoh17/simple-e2e/framework/util"
 	"github.com/stretchr/testify/assert"
 )
@@ -81,7 +79,7 @@ func TestGlobalVariablesMultiStage(t *testing.T) {
 }
 
 func unmarshalYaml(fileName string, t *testing.T) Procedure {
-	SetTestFilesRoot()
+	internal.SetTestFilesRoot()
 	var procedure Procedure
 
 	body, err := ioutil.ReadFile(fmt.Sprintf("%s/examples/%s.yaml", os.Getenv(util.TestDirEnv), fileName))
@@ -95,13 +93,4 @@ func unmarshalYaml(fileName string, t *testing.T) Procedure {
 	}
 
 	return procedure
-}
-
-func SetTestFilesRoot() {
-	// If not in container, set as the path to the 'project's root/tests'
-	if os.Getenv(util.TestDirEnv) == "" {
-		_, b, _, _ := runtime.Caller(0)
-		d := path.Join(path.Dir(b))
-		os.Setenv(util.TestDirEnv, fmt.Sprintf("%s/../tests", filepath.Dir(d)))
-	}
 }
