+++ framework/operations/stepManager_test.go
@@ -2,9 +2,9 @@ package operations
 
 import (
 	"fmt"
-	"os"
 	"testing"
 
+	"github.com/julianGoh17/simple-e2e/framework/internal"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -133,17 +133,5 @@ func TestAddingDuplicateLiteralTestSteps(t *testing.T) {
 }
 
 func TestMain(m *testing.M) {
-	// call flag.Parse() here if TestMain uses flags
-	rc := m.Run()
-
-	// rc 0 means we've passed,
-	// and CoverMode will be non empty if run with -cover
-	if rc == 0 && testing.CoverMode() != "" {
-		c := testing.Coverage()
-		if c < 0.85 {
-			fmt.Println("Tests passed but coverage failed at", c)
-			rc = -1
-		}
-	}
-	os.Exit(rc)
+	internal.TestCoverageReaches85Percent(m)
 }
